backend/requirements.txt
# Python Backend Requirements for CineCritique AI
# Target: Python 3.12+

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.10.0
pydantic-settings==2.6.1

# Video Processing
opencv-python==4.10.0.84
librosa==0.10.2.post1
ffmpeg-python==0.2.0
numpy==1.26.4
soundfile==0.12.1

# Local LLM Support
llama-cpp-python==0.3.2
transformers==4.46.3
torch==2.5.1
sentence-transformers==3.3.1

# Caching & Storage
redis==5.2.0
aioredis==2.0.1
diskcache==5.6.3

# AI SDK
google-generativeai==0.8.3

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0
pillow==11.0.0
httpx==0.28.1

# Development
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.8.4
